--- website/lib/app/header.stories.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-import Header from "./header"
-
-export default {
-  title: 'App/Header',
-  component: Header,
-};
-
-const Template = (args) => <Header />
-
-export const Primary = Template.bind({});
-Primary.args = {
-  primary: true,
-  label: 'App',
-};
