--- website/lib/changelog/index.js
@@ -1,8 +1,0 @@
-
-import Changelog from "./components/changelog"
-import processChangelog from "./nextjs/changelog"
-
-export {
-    Changelog,
-    processChangelog
-}
