--- website/lib/changelog/styles/changelog.css
@@ -1,70 +1,0 @@
-
-.changelog {
-    font-size: 1.125rem;
-    line-height: 1.7777778;
-}
-
-
-.changelog ul {
-    margin-top: 1.3333333em;
-    margin-bottom: 1.3333333em;
-}
-
-.changelog ul > li {
-    position: relative;
-    padding-left: 1.75em;
-}
-
-.changelog ul > li::before {
-    width: 0.3333333em;
-    height: 0.3333333em;
-    top: calc(0.8888889em - 0.1666667em);
-    left: 0.2222222em;
-}
-
-.changelog ul > li::before {
-    content: "";
-    position: absolute;
-    background-color: #d1d5db;
-    border-radius: 50%;
-    width: 0.375em;
-    height: 0.375em;
-    top: calc(0.875em - 0.1875em);
-    left: 0.25em;
-}
-
-.changelog li {
-    margin-top: 0.6666667em;
-    margin-bottom: 0.6666667em;
-}
-
-.changelog h1 {
-    font-size: 2.6666667em;
-    margin-top: 0;
-    margin-bottom: 0.8333333em;
-    line-height: 1;
-    font-weight: 800;
-}
-
-.changelog h2 {
-    font-size: 1.6666667em;
-    margin-top: 40px;
-    margin-bottom: 24px;
-    line-height: 1.3333333;
-    font-weight: 700;
-}
-
-.changelog h3 {
-    font-size: 1.3333333em;
-    margin-bottom: 8px;
-    line-height: 1.5;
-    font-weight: 600;
-}
-
-.changelog h2 + * {
-    margin-top: 0;
-}
-
-.changelog pre {
-    margin: 1.5rem 0;
-}
