--- website/lib/docs/index.js
@@ -1,9 +1,0 @@
-
-import { getSidebarSlugs, getData } from './nextjs/content'
-import Docs from './components/docs'
-
-export {
-    getSidebarSlugs,
-    getData,
-    Docs
-}
