--- website/lib/landing/index.js
@@ -1,12 +1,0 @@
-
-import Cta from "./cta"
-import Hero from "./hero"
-import Features from "./features"
-import UseCases from "./use-cases"
-
-export default {
-    Cta,
-    Hero,
-    Features,
-    UseCases,
-}
