--- website/lib/landing/use-cases.js
@@ -1,38 +1,0 @@
-
-/* This example requires Tailwind CSS v2.0+ */
-
-  export default function Example({features}) {
-    return (
-      <div className="relative bg-use-cases bg-white2 py-16 sm:py-24">
-        <div className="mx-auto max-w-md px-4 text-center sm:max-w-3xl sm:px-6 lg:px-8 lg:max-w-7xl">
-          <p className="mt-2 text-3xl font-extrabold text-gray-900 tracking-tight sm:text-4xl">
-            Use cases
-          </p>
-          {
-              /*
-          <p className="mt-5 max-w-prose mx-auto text-xl text-gray-500">
-            Phasellus lorem quam molestie id quisque diam aenean nulla in. Accumsan in quis quis nunc, ullamcorper
-            malesuada. Eleifend condimentum id viverra nulla.
-          </p>
-          */ }
-          <div className="mt-12">
-            <div className="grid grid-cols-1 gap-8 sm:grid-cols-2 lg:grid-cols-3">
-              {features.map((feature) => (
-                <div key={feature.name} className="text-left border border-black shadow-use-cases bg-white">
-                  <div className="px-6 pb-8">
-                    <div className="">
-                      <h3 className="mt-8 text-lg font-medium text-gray-900 tracking-tight">{feature.name}</h3>
-                      <p className="mt-5 text-base text-gray-500">
-                        {feature.text}
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-      </div>
-    )
-  }
-
