+++ website/package.json
@@ -14,7 +14,12 @@
     "@mapbox/rehype-prism": "^0.5.0",
     "@sindresorhus/slugify": "^1.1.0",
     "@svgr/webpack": "^5.5.0",
-    "@teseraio/cookie-consent-manager": "1.0.3",
+    "@teseraio/cookie-consent-manager": "1.0.4",
+    "@teseraio/oss-react-app": "^0.0.3",
+    "@teseraio/oss-react-changelog": "^0.0.4",
+    "@teseraio/oss-react-community": "0.0.2",
+    "@teseraio/oss-react-docs": "0.0.5",
+    "@teseraio/oss-react-landing": "0.0.5",
     "@teseraio/tesera-oss": "1.0.1",
     "autoprefixer": "^9",
     "clsx": "^1.1.1",
