+++ website/pages/changelog/index.js
@@ -1,8 +1,16 @@
 
-import {Changelog} from "@teseraio/tesera-oss"
+import Changelog from "@teseraio/oss-react-changelog"
+import Head from 'next/head'
 
 export default function Index(data) {
-    return <Changelog.Changelog {...data} />
+    return (
+        <div>
+            <Head>
+                <title key="title">Changelog | Ensemble</title>
+            </Head>
+            <Changelog.Changelog {...data} />
+        </div>
+    )
 }
 
 export async function getStaticProps() {
