+++ website/pages/docs/[...page].js
@@ -1,14 +1,20 @@
 
 import sidebarContent from "../../data/sidebar-docs.json"
-
-import { Docs } from "@teseraio/tesera-oss"
-//import { getSidebarSlugs, getData, Docs } from '../../lib/docs'
+import Docs from "@teseraio/oss-react-docs"
+import Head from 'next/head'
 
 export default function Post({postData}) {
-    return <Docs.Docs postData={postData} sidebar={sidebarContent} />
+    return (
+        <div>
+            <Head>
+                <title key="title">{`Docs - ${postData.title} | Ensemble`}</title>
+            </Head>
+            <Docs.Docs postData={postData} sidebar={sidebarContent} />
+        </div>
+    )
 }
 
-const docsPrefix = "/docs/"
+const docsPrefix = "/docs"
 
 export async function getStaticProps({ params }) {
     console.log("-- params --")
@@ -20,6 +26,9 @@ export async function getStaticProps({ params }) {
 export async function getStaticPaths() {
     const xx =  Docs.getSidebarSlugs(docsPrefix)
 
+    console.log("-- xxxxxx --")
+    console.log(xx)
+    
     console.log(JSON.stringify(xx))
     return xx
 }
