+++ website/pages/docs/index.js
@@ -1,12 +1,12 @@
 
-import {Docs} from '../../lib/docs'
+import Docs from "@teseraio/oss-react-docs"
 import sidebarContent from "../../data/sidebar-docs.json"
 
 export default function FirstPost({sidebar}) {
   const postData = {
     title: 'Documentation',
   }
-  return <Docs main={Index} postData={postData} sidebar={sidebarContent}/>
+  return <Docs.Docs main={Index} postData={postData} sidebar={sidebarContent}/>
 }
 
 const Index = () => (
