+++ website/styles/index.css
@@ -40,207 +40,6 @@
   font-display: swap;
 }
 
-/*
-.btn-blue {
-  @apply bg-blue-500 text-white font-bold py-2 px-4 rounded;
-}
-
-.anchor {
-  display: inline-block;
-  opacity: 0;
-  transition: opacity .2s ease;
-  order: 1;
-  position: relative;
-  left: -.7em;
-  margin-right: -.5em;
-}
-
-.anchor:hover{
-  opacity: 1;
-}
-
-.some {
-  display: block;
-  position: relative;
-  top: -4rem;
-  visibility: hidden;
-}
-
-.grid-style::after {
-  background: linear-gradient(90deg, black 1px, transparent 1%) 0px 0;
-  background-size: 20%;
-  border-right-width: 1px;
-}
-
-h1 {
-  font-size: 2.5rem;
-}
-
-h2 {
-  font-size: 2rem;
-}
-
-h3 {
-  font-size: 1.5rem;
-}
-
-#nav a {
-  padding: 7px 0 7px 5px;
-  display: block;
-  transition: color .2s ease;
-  cursor: pointer;
-}
-
-.rotate {
-  transform: rotate(90deg);
-}
-
-#app a {
-  color: #6B1FFF;
-  position: relative;
-  transition: color .3s;
-  text-decoration: none;
-}
-
-#app a:hover {
-  text-decoration: underline;
-}
-
-.reason-image svg {
-  width: 100%;
-  height: auto;
-  display: block;
-  margin: 0 auto;
-}
-
-#app strong {
-  background: #f4f4f4;
-  font-size: 1.20rem;
-  font-weight: normal;
-  padding: 2px 7px;
-  border-color: #ccc;
-  border-width: 1px;
-}
-
-.xxx {
-  border-top: 2px solid #e5eaed;
-}
-
-.reason p {
-  margin-bottom: 15px;
-}
-
-.use-case-shadow {
-  -webkit-box-shadow: 7px 7px 0px 0px rgba(50, 50, 50, 0.11);
-  -moz-box-shadow:    7px 7px 0px 0px rgba(50, 50, 50, 0.11);
-  box-shadow:         7px 7px 0px 0px rgba(50, 50, 50, 0.11);
-}
-
-.why li::before {
-    content: '';
-    display: inline-block;
-    height: 1em;
-    width: 1em;
-    background-image: url(/tick.svg);
-    background-size: contain;
-    background-repeat: no-repeat;
-    margin-right:0.5em;
-    top: 5px;
-    position: relative;
-}
-
-.github-icon {
-  align-items: center;
-  padding: 9px 10px;
-}
-
-.github-stars {
-  align-items: center;
-  height: 100%;
-  border-left: 1px solid rgba(29,30,35,.2);
-  padding: 6px 12px 7px 10px;
-  transition: border-color .2s;
-  z-index: 1;
-  font-family: var(--font-body);
-  font-size: .938rem;
-  letter-spacing: .01em;
-  line-height: 1.667em;
-  font-weight: 600;
-}
-
-.use-cases {
-  background-image: url(/texture-squares.svg);
-}
-
-.content ul { 
-  list-style-type: disc; 
-  list-style-position: inside; 
-}
-
-.content ul {
-  margin-top: 8px;
-}
-
-.content pre {
-  margin-top: 15px;
-}
-
-.content p { 
-  margin-top: 15px;
-}
-
-.content h3 {
-  margin-top: 15px;
-}
-
-.content hr {
-  margin-top: 25px;
-}
-
-.content h2 {
-  font-weight: bold;
-  margin-top:20px;
-}
-*/
-
-ul.experiences{
-  list-style: none;
-  margin:0;
-  padding:0;
-  position: relative;
-}
-
-ul.experiences:before{
-  content:"";
-  display: inline-block;
-  width: 2px;
-  background: red;
-  position: absolute;
-  left:25px;
-  top:5px;
-  height: calc(100% - 10px );
-}
-
-/*
-ul.experiences li{
-    position: relative;
-}
-
-ul.experiences li:before{
-      content:"";
-      display: inline-block;
-      width: 8px;
-      height: 8px;
-      background: red;
-      position: absolute;
-      left: -18px;
-      top: 14px;
-      border-radius: 10px;
-}
-*/
-
-@import 'prism.css';
-
 /* Start purging... */
 @tailwind utilities;
 /* Stop purging. */
