+++ website/styles/use-cases.css
@@ -1,0 +1,4 @@
+
+.use-cases {
+    background-image: url('/texture-squares.svg')
+}
